Modifications for Skills to be used by Wifeys and Enemies
