+++ app/src/main/java/ultrapoulet/wifeygame/battle/BattleCharacter.java
@@ -1,189 +1,19 @@
 package ultrapoulet.wifeygame.battle;
 
-import ultrapoulet.androidgame.framework.Image;
-import ultrapoulet.wifeygame.character.Weapon;
-import ultrapoulet.wifeygame.skills.SkillList;
-
 /**
- * Created by John on 3/5/2016.
+ * Created by John on 7/13/2016.
  */
-public class BattleCharacter {
-
-    private String name;
-    private int maxHP;
-    private int currentHP;
-    private int numHits;
-    private Weapon weapon;
-    private SkillList skills;
-    private int strength;
-    private int magic;
-
-    private Image image;
-
-    private boolean isDefending = false;
-
-    //Section for skill flags
-
-    public BattleCharacter(String name, Weapon weapon, int strength, int magic, Image image, SkillList skills){
-        this.name = name;
-        this.maxHP = calculateHP(strength);
-        this.currentHP = this.maxHP;
-        this.weapon = weapon;
-        this.numHits = weapon.getNumHits();
-        this.strength = strength;
-        this.magic = magic;
-        this.image = image;
-        this.skills = skills;
-        skills.setOwner(this);
-    }
-
-    public static int calculateHP(int strength){
-        return 10 * strength;
-    }
-
-    public String getName(){
-        return this.name;
-    }
-
-    public Weapon getWeapon() { return this.weapon; }
-
-    public int getNumHits(){
-        return this.numHits;
-    }
-
-    public int getMaxHP(){
-        return this.maxHP;
-    }
-
-    public int getCurrentHP(){
-        return this.currentHP;
-    }
-
-    public int getStrength(){
-        return this.strength;
-    }
-
-    public int getMagic(){
-        return this.magic;
-    }
-
-    public Image getImage(){
-        return this.image;
-    }
-
-    //For testing only
-    public void setCurrentHP(int hp){
-        this.currentHP = hp;
-    }
-
-    public void battleStart(){
-        this.currentHP = this.maxHP;
-        this.isDefending = false;
-    }
-
-    public void turnStart(){
-        //Function for things that happen at start of turn
-        this.isDefending = false;
-    }
-
-    public int PowerAttackDamage(){
-        int baseDamage = this.strength * 5;
-        //Do checks on skills to determine bonus damage
-        //int modDamage = (int) (baseDamage * skills.physicalDamagePercentage(enemy));
-        int modDamage = baseDamage + (int) ((baseDamage / 10) * Math.random());
-        return modDamage;
-    }
-
-    public int ComboAttackDamage(){
-        int powerDamage = PowerAttackDamage();
-        int totalHits = this.numHits + this.skills.getBonusHits();
-        if(totalHits > 10){
-            totalHits =  10;
-        }
-        int divider = (90 - 4*(totalHits - 2)) / totalHits;
-        return (powerDamage * divider) / 100;
-    }
-
-    public int MagicAttackDamage(){
-        int baseDamage = this.magic * 5;
-        //Do checks on skills to determine bonus damage
-        //int modDamage = (int) (baseDamage * skills.magicalDamagePercentage(enemy));
-        int modDamage = baseDamage + (int) ((baseDamage / 10) * Math.random());
-        return modDamage;
-    }
-
-    public int HealAmount(){
-        int baseHeal = this.magic * 2;
-        //Do checks on skills to determine bonus healing
-        //int modHeal = (int) (baseHeal * skills.healPercentage(member);
-        int modHeal = baseHeal + (int) ((baseHeal / 10) * Math.random());
-        return baseHeal;
-    }
-
-    public int SpecialDamage(){
-        int baseDamage = this.strength * 4 + this.magic * 4;
-        //Do checks on skills to determine bonus damage
-        //int modDamage = (int) (baseDamage * skills.specialDamagePercentage(enemy));
-        int modDamage = baseDamage + (int) ((baseDamage / 10) * Math.random());
-        return modDamage;
-
-    }
-
-    public void Defend(){
-        this.isDefending = true;
-    }
-
-    public int takePhysicalDamage(int damage){
-        //Check skills for anything to reduce damage taken
-        int displayDamage = damage;
-        if(this.isDefending){ displayDamage = displayDamage/2; }
-        //displayDamage = (int) (displayDamage * skills.receivePhysicalAttackPercentage(enemy));
-        //skills.onDamageReceived(displayDamage);
-        this.currentHP = this.currentHP - displayDamage;
-        if(this.currentHP <= 0){
-            this.currentHP = 0;
-        }
-        //Check if there's a skill to prevent lethal damage
-        return displayDamage;
-    }
-
-    public int takeMagicalDamage(int damage){
-        //Check skills for anything to reduce damage taken
-        int displayDamage = damage;
-        if(this.isDefending){ displayDamage = displayDamage/2; }
-        //displayDamage = (int) (displayDamage * skills.receiveMagicalAttackPercentage(enemy));
-        //skills.onDamageReceived(displayDamage);
-        this.currentHP = this.currentHP - displayDamage;
-        if(this.currentHP <= 0){
-            this.currentHP = 0;
-        }
-        return displayDamage;
-    }
-
-
-    public int takeSpecialDamage(int damage){
-        //Check skills for anything to reduce damage taken
-        int displayDamage = damage;
-        if(this.isDefending){ displayDamage = displayDamage/2; }
-        //displayDamage = (int) (displayDamage * skills.receiveSpecialAttackPercentage(enemy));
-        //skills.onDamageReceived(displayDamage);
-        this.currentHP = this.currentHP - displayDamage;
-        if(this.currentHP <= 0){
-            this.currentHP = 0;
-        }
-        return displayDamage;
-    }
-
-
-    public int healDamage(int heal){
-        //Check skills for anything to increase healing
-        int displayHeal = heal;
-        this.currentHP = this.currentHP + displayHeal;
-        //displayHeal = (int) (displayHeal * skills.receiveHealPercentage(healer);
-        if(this.currentHP > this.maxHP){
-            displayHeal = displayHeal - (this.currentHP - this.maxHP);
-            this.currentHP = this.maxHP;
-        }
-        return displayHeal;
-    }
+public interface BattleCharacter {
+    public int getCurrentHP();
+    public int getMaxHP();
+    public void battleStart();
+    public int PowerAttackDamage();
+    public int ComboAttackDamage();
+    public int MagicAttackDamage();
+    public int HealAmount();
+    public int SpecialAttackDamage();
+    public int takePhysicalDamage(int damage);
+    public int takeMagicalDamage(int damage);
+    public int takeSpecialDamage(int damage);
+    public int healDamage(int heal);
 }
