+++ app/src/main/java/ultrapoulet/wifeygame/skills/SkillList.java
@@ -3,7 +3,6 @@ package ultrapoulet.wifeygame.skills;
 import java.util.ArrayList;
 
 import ultrapoulet.wifeygame.battle.BattleCharacter;
-import ultrapoulet.wifeygame.battle.BattleEnemy;
 
 /**
  * Created by John on 7/11/2016.
@@ -49,7 +48,7 @@ public class SkillList {
         }
     }
 
-    public double physicalAttackPercentage(BattleEnemy enemy){
+    public double physicalAttackPercentage(BattleCharacter enemy){
         //The amount to increase physical damage.
         //Some skill might need to see the state of the BattleEnemy, so it is included
         //Returned multiplier is additive. It will start at 1.0, and add/subtract values from
@@ -71,7 +70,7 @@ public class SkillList {
         return 0;
     }
 
-    public double magicalAttackPercentage(BattleEnemy enemy){
+    public double magicalAttackPercentage(BattleCharacter enemy){
         //The amount to increase magical damage.
         //Some skills might need to see the state of the BattleEnemy, so it is included
         //Returned multiplier is additive. It will start at 1.0, and add/subtract values
@@ -83,7 +82,7 @@ public class SkillList {
         return multiplier;
     }
 
-    public double specialAttackPercentage(BattleEnemy enemy){
+    public double specialAttackPercentage(BattleCharacter enemy){
         //The amount to increase special damage.
         //Some skills might need to see the state of the BattleEnemy, so it is included
         //Returned multiplier is additive. It will start at 1.0, and add/subtract values
@@ -105,7 +104,7 @@ public class SkillList {
         return multiplier;
     }
 
-    public double receivePhysicalAttackPercentage(BattleEnemy enemy){
+    public double receivePhysicalAttackPercentage(BattleCharacter enemy){
         //The amount to decrease physical damage.
         //This will be capped, so as not to hit 0 or below
         double multiplier = 1.0;
@@ -118,7 +117,7 @@ public class SkillList {
         return multiplier;
     }
 
-    public double receiveMagicalAttackPercentage(BattleEnemy enemy){
+    public double receiveMagicalAttackPercentage(BattleCharacter enemy){
         //The amount to decrease magical damage.
         //This will be capped, so as not to hit 0 or below
         double multiplier = 1.0;
@@ -131,7 +130,7 @@ public class SkillList {
         return multiplier;
     }
 
-    public double receiveSpecialAttackPercentage(BattleEnemy enemy){
+    public double receiveSpecialAttackPercentage(BattleCharacter enemy){
         //The amount to decrease special damage.
         //This will be capped, so as not to hit 0 or below
         double multiplier = 1.0;
@@ -162,7 +161,7 @@ public class SkillList {
         }
     }
 
-    public void onEnemyDefeat(BattleEnemy enemy){
+    public void onEnemyDefeat(BattleCharacter enemy){
         //Do things for when an enemy is defeated
         for(int i = 0; i < skills.size(); i++){
             skills.get(i).onEnemyDefeat(enemy);
